rasterize-pdf + offline mode

+++ Integrations/rasterize/rasterize.py
@@ -4,6 +4,12 @@ from CommonServerUserPython import *
 
 from selenium import webdriver
 from selenium.common.exceptions import NoSuchElementException, InvalidArgumentException
+from PyPDF2 import PdfFileReader
+from pdf2image import convert_from_path
+import numpy as np
+from PIL import Image
+import tempfile
+from io import BytesIO
 import sys
 import base64
 
@@ -29,11 +35,11 @@ def check_response(driver):
         return_error(EMPTY_RESPONSE_ERROR_MSG) if WITH_ERRORS else return_warning(EMPTY_RESPONSE_ERROR_MSG, exit=True)
 
 
-def init_driver():
+def init_driver(offline_mode=False):
     """
     Creates headless Google Chrome Web Driver
     """
-    demisto.debug('Creating chrome driver')
+    demisto.debug(f'Creating chrome driver. Mode: {"OFFLINE" if offline_mode else "ONLINE"}')
     try:
         with open('log.txt', 'w') as log:
             sys.stdout = log
@@ -48,9 +54,11 @@ def init_driver():
             chrome_options.add_argument('--ignore-certificate-errors')
 
             driver = webdriver.Chrome(options=chrome_options)
+            if offline_mode:
+                driver.set_network_conditions(offline=True, latency=5, throughput=500 * 1024)
 
-            # remove log
-            os.remove(os.path.realpath(log.name))
+                # remove log
+                os.remove(os.path.realpath(log.name))
 
     except Exception as ex:
         return_error(str(ex))
@@ -61,18 +69,19 @@ def init_driver():
     return driver
 
 
-def rasterize(path: str, width: int, height: int, r_type='png'):
+def rasterize(path: str, width: int, height: int, r_type='png', offline_mode=False):
     """
     Capturing a snapshot of a path (url/file), using Chrome Driver
+    :param offline_mode: when set to True, will block any outgoing communication
     :param path: file path, or website url
     :param width: desired snapshot width in pixels
     :param height: desired snapshot height in pixels
     :param r_type: result type: .png/.pdf
     """
-    driver = init_driver()
+    driver = init_driver(offline_mode)
 
     try:
-        demisto.debug('Navigating to path')
+        demisto.debug(f'Navigating to path. Mode: {"OFFLINE" if offline_mode else "ONLINE"}')
 
         driver.get(path)
         driver.implicitly_wait(5)
@@ -135,6 +144,44 @@ def get_pdf(driver, width: int, height: int):
     return data
 
 
+def convert_pdf_to_jpeg(path: str, max_pages: int, password: str, horizontal: bool = False):
+    demisto.debug(f'Loading file at Path: {path}')
+    inputpdf = PdfFileReader(open(path, "rb"))
+    pages = min(max_pages, inputpdf.numPages)
+
+    with tempfile.TemporaryDirectory() as output_folder:
+        demisto.debug('Converting PDF')
+        convert_from_path(
+            pdf_path=path,
+            fmt='jpeg',
+            first_page=1,
+            last_page=pages,
+            output_folder=output_folder,
+            userpw=password,
+            output_file='converted_pdf_'
+        )
+        demisto.debug('Converting PDF - COMPLETED')
+
+        demisto.debug('Combining all pages')
+        images = []
+        for i in sorted(os.listdir(output_folder)):
+            if os.path.isfile(os.path.join(output_folder, i)) and 'converted_pdf_' in i:
+                images.append(Image.open(os.path.join(output_folder, i)))
+        min_shape = min([(np.sum(i.size), i.size) for i in images])[1]
+
+        if horizontal:
+            imgs_comb = np.hstack([np.asarray(i.resize(min_shape)) for i in images])
+        else:
+            imgs_comb = np.vstack([np.asarray(i.resize(min_shape)) for i in images])
+
+        imgs_comb = Image.fromarray(imgs_comb)
+        output = BytesIO()
+        imgs_comb.save(output, 'JPEG')
+        demisto.debug('Combining all pages - COMPLETED')
+
+        return output.getvalue()
+
+
 def rasterize_command():
     url = demisto.getArg('url')
     w = demisto.args().get('width', DEFAULT_W)
@@ -179,6 +226,7 @@ def rasterize_email_command():
     html_body = demisto.args().get('htmlBody')
     w = demisto.args().get('width', DEFAULT_W)
     h = demisto.args().get('height', DEFAULT_H)
+    offline = demisto.args().get('offline', 'false') == 'true'
     r_type = demisto.args().get('type', 'png')
 
     filename = f'email.{"pdf" if r_type.lower() == "pdf" else "png"}'  # type: ignore
@@ -186,12 +234,31 @@ def rasterize_email_command():
         f.write(f'<html style="background:white";>{html_body}</html>')
     path = f'file://{os.path.realpath(f.name)}'
 
-    output = rasterize(path=path, r_type=r_type, width=w, height=h)
+    output = rasterize(path=path, r_type=r_type, width=w, height=h, offline_mode=offline)
     file = fileResult(filename=filename, data=output)
     if r_type == 'png':
         file['Type'] = entryTypes['image']
 
     demisto.results(file)
+
+
+def rasterize_pdf_command():
+    entry_id = demisto.args().get('EntryID')
+    password = demisto.args().get('pdfPassword')
+    max_pages = int(demisto.args().get('maxPages', 30))
+    horizontal = demisto.args().get('horizontal', 'false') == 'true'
+
+    file_path = demisto.getFilePath(entry_id).get('path')
+
+    filename = 'image.jpeg'  # type: ignore
+
+    with open(file_path, 'rb') as f:
+        output = convert_pdf_to_jpeg(path=os.path.realpath(f.name), max_pages=max_pages, password=password,
+                                     horizontal=horizontal)
+        file = fileResult(filename=filename, data=output)
+        file['Type'] = entryTypes['image']
+
+        demisto.results(file)
 
 
 def test():
@@ -217,6 +284,9 @@ def main():
 
         elif demisto.command() == 'rasterize-email':
             rasterize_email_command()
+
+        elif demisto.command() == 'rasterize-pdf':
+            rasterize_pdf_command()
 
         elif demisto.command() == 'rasterize':
             rasterize_command()
